+++ apps/nextjs/src/app/aila/page-contents.tsx
@@ -2,12 +2,13 @@
 
 import React from "react";
 
+import { MathJaxContext } from "better-react-mathjax";
+
 import { AiSdk } from "@/components/AppComponents/Chat/AiSdk";
 import { Chat } from "@/components/AppComponents/Chat/Chat/chat";
 import LessonPlanDisplay from "@/components/AppComponents/Chat/chat-lessonPlanDisplay";
 import Layout from "@/components/AppComponents/Layout";
 import { useReactScan } from "@/hooks/useReactScan";
-import LessonPlanTrackingProvider from "@/lib/analytics/lessonPlanTrackingContext";
 import { AilaStoresProvider } from "@/stores/AilaStoresProvider";
 
 const ChatPageContents = ({ id }: { readonly id: string }) => {
@@ -15,12 +16,12 @@ const ChatPageContents = ({ id }: { readonly id: string }) => {
 
   return (
     <Layout>
-      <LessonPlanTrackingProvider chatId={id}>
+      <MathJaxContext>
         <AilaStoresProvider id={id}>
           <AiSdk id={id} />
           <Chat />
         </AilaStoresProvider>
-      </LessonPlanTrackingProvider>
+      </MathJaxContext>
     </Layout>
   );
 };