+++ packages/logger/index.ts
@@ -16,11 +16,13 @@ debugBase.log = console.log.bind(console);
 
 export type LoggerKey =
   | "admin"
+  | "additional-materials"
   | "aila"
   | "aila:analytics"
   | "aila:categorisation"
   | "aila:chat"
   | "aila:errors"
+  | "aila:experimental-patches"
   | "aila:lesson"
   | "aila:llm"
   | "aila:moderation"
@@ -28,13 +30,13 @@ export type LoggerKey =
   | "aila:persistence"
   | "aila:prompt"
   | "aila:protocol"
-  | "aila:stream"
-  | "aila:rag"
-  | "aila:testing"
   | "aila:quiz"
-  | "aila:experimental-patches"
+  | "aila:rag"
+  | "aila:stream"
+  | "aila:testing"
   | "aila:threat"
   | "analytics"
+  | "analytics:lesson:store"
   | "app"
   | "auth"
   | "chat"
@@ -74,11 +76,6 @@ export type LoggerKey =
   | "ui:performance"
   | "webhooks";
 
-const errorLogger =
-  typeof window === "undefined"
-    ? structuredLogger.error.bind(structuredLogger)
-    : browserLogger.error.bind(browserLogger);
-
 /**
  * The AI logger uses namespaces so that we can selectively toggle noisy logs.
  * Logs are selected with the DEBUG environment variable.
@@ -104,7 +101,7 @@ export function aiLogger(childKey: LoggerKey) {
   return {
     info: debugLogger,
     warn: debugLogger,
-    error: errorLogger.bind(structuredLogger),
+    error: console.error,
     table: tableLogger,
   };
 }